Movie 관련 서브 클래스 생성
