+++ src/main/java/com/videorental/Movie.java
@@ -1,6 +1,8 @@
 package com.videorental;
 
-public class Movie {
+import javax.swing.text.AsyncBoxView;
+
+public abstract class Movie {
     public static final int REGULAR = 1;
     public static final int NEW_RELEASE = 2;
     public static final int CHILDREN = 3;
@@ -26,27 +28,7 @@ public class Movie {
     }
 
     // move instance method + refactoring
-    double getChargeFor(int daysRented, Rental rental) {
-        double thisAmount = 0;
-        switch (getPriceCode()) {
-            case REGULAR:
-                thisAmount += 2;
-                if (daysRented > 2)
-                    thisAmount += (daysRented - 2) * 1.5;
-                break;
-
-            case NEW_RELEASE:
-                thisAmount += daysRented * 3;
-                break;
-
-            case CHILDREN:
-                thisAmount += 1.5;
-                if (daysRented > 3)
-                    thisAmount += (daysRented - 3) * 1.5;
-                break;
-        }
-        return thisAmount;
-    }
+    abstract double getChargeFor(int daysRented, Rental rental);
 
     int getFrequentRentalPointsFor(int daysRented) {
         // add bonus for a two day new release rental
